requirements.txt
amqp==2.2.2
retrying==1.3.3
ldap3==2.4.1
Django==3.2
#py2neo==3.1.2
py2neo==2021.1.2
requests
pycrypto==2.6.1
djangorestframework==3.12.0
PyYAML==5.4
django_nose==1.4.5
nose==1.3.7
pyasn1==0.3.3
pyasn1-modules==0.1.1
graypy==0.2.14
pycryptodome==3.15.0